+++ components/WhenYouDoEachActivity.tsx
