+++ components/insights/GroupBySelector.tsx
