+++ components/insights/NumDaysSelector.tsx
